requirements.txt
###### Requirements without Version Specifiers ######
Jinja2
aiofiles

###### Requirements with Version Specifiers ######`
numpy>=1.17.4
Pillow>=8.4.0
pydantic>=1.8.2
torch>=1.10.0+cu113
torchvision>=0.11.1+cu113
facenet_pytorch>=2.5.2
faiss_gpu>=1.7.1.post2
fastapi>=0.70.0
redis>=4.0.2
mariadb>=1.0.8
sqlalchemy>=1.4.28